Fixed bus line display
Fixed the displaying of the bus line. Now the bus line is always
displayed either there is wait time information or not.

Changed the error display format: Connection error will be shown now
with a small dot over the waiting time and a not updated wait time will
be signaled with the semicolon between the line and the wait time.